async_diffusion_policy-master/async_dp/pyproject.toml
[project]
name = "async-dp"
version = "0.1.0"
description = "Asynchronous Diffusion Policy for robot control"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "diffusers>=0.27.0",
    "einops>=0.8.0",
    "accelerate>=0.28.0",
    "numpy>=1.26.0,<2.0.0",
    "h5py>=3.10.0",
    "scipy>=1.12.0",
    "matplotlib>=3.8.0",
    "tqdm>=4.66.0",
    "typer>=0.9.0",
    "pyyaml>=6.0.1",
    "modern-robotics>=1.1.1",
    "dynamixel-sdk>=3.7.31",
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
ros = [
    "catkin_pkg",
    "rospkg",
    "empy",
]
grpc = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "config"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
